Split project into mhook and mhook-test project.
The mhook project creates the mhook library, and creates a directory for
public headers.
The mhook-test project consumes outputs of mhook, and is used for testing
only.
The purpose of this change is to create an mhook library that can be
consumed as a library. It still leaves an option to use mhook code
directly.
