--- mhook-test.sln
@@ -1,29 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mhook-test", "mhook-test.vcxproj", "{0E055CAF-C68B-42CB-A302-F775CA5A917F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|Win32.ActiveCfg = Debug|Win32
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|Win32.Build.0 = Debug|Win32
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x64.ActiveCfg = Debug|x64
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x64.Build.0 = Debug|x64
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|Win32.ActiveCfg = Release|Win32
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|Win32.Build.0 = Release|Win32
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x64.ActiveCfg = Release|x64
-		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(DPCodeReviewSolutionGUID) = preSolution
-		DPCodeReviewSolutionGUID = {00000000-0000-0000-0000-000000000000}
-	EndGlobalSection
-EndGlobal
