+++ mhook.sln
@@ -1,0 +1,41 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mhook", "mhook\mhook.vcxproj", "{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "mhook-test", "mhook-test\mhook-test.vcxproj", "{0E055CAF-C68B-42CB-A302-F775CA5A917F}"
+	ProjectSection(ProjectDependencies) = postProject
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7} = {DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Debug|x64.ActiveCfg = Debug|x64
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Debug|x64.Build.0 = Debug|x64
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Debug|x86.ActiveCfg = Debug|Win32
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Debug|x86.Build.0 = Debug|Win32
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Release|x64.ActiveCfg = Release|x64
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Release|x64.Build.0 = Release|x64
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Release|x86.ActiveCfg = Release|Win32
+		{DFBE49D1-548A-4CF9-B9AC-705642E0CFF7}.Release|x86.Build.0 = Release|Win32
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x64.ActiveCfg = Debug|x64
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x64.Build.0 = Debug|x64
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x86.ActiveCfg = Debug|Win32
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Debug|x86.Build.0 = Debug|Win32
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x64.ActiveCfg = Release|x64
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x64.Build.0 = Release|x64
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x86.ActiveCfg = Release|Win32
+		{0E055CAF-C68B-42CB-A302-F775CA5A917F}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
