+++ mhook/src/disasm-lib/cpu.c
