+++ mhook/src/disasm-lib/disasm.c
