+++ mhook/src/disasm-lib/misc.c
