+++ mhook/src/disasm-lib/misc.h
